fixed small mistakes, renamed some styles for clarity
