+++ src/components/Header/Header/styles.tsx
@@ -1,15 +1,14 @@
 import styled from 'styled-components'
 
-export const HeaderWrapper = styled.div<{ HeaderBackground: string }>`
+export const HeaderWrapper = styled.div<{ headerBackground: string }>`
   display: flex;
-  flex-direction: row;
   background-color: #2819f5;
   justify-content: space-between;
   margin: 0;
-  padding: 2px 50px;
+  padding: 2px auto;
   height: auto;
   color: white;
-  background-image: url(${(props) => props.HeaderBackground});
+  background-image: url(${(props) => props.headerBackground});
   background-blend-mode: multiply;
   background-size: cover;
 `
