+++ src/components/Header/HeaderLeft/styles.tsx
@@ -1,14 +1,14 @@
 import styled from 'styled-components'
 
 export const HeaderLeftWrapper = styled.div`
-  font-size: 28px;
-  margin-left: 50px auto;
+  margin-left: 50px;
   padding: 8px;
   text-transform: uppercase;
   font-weight: bold;
   display: flex;
   align-items: center;
 `
-export const H1 = styled.h1`
+export const NameTitle = styled.h1`
   margin: 0 auto;
+  font-size: 48px;
 `
