+++ src/components/body/AboutMe/index.tsx
@@ -3,7 +3,7 @@ import { Icon } from '../../general/Icon'
 import frisbeeLogo from '../../../thumbnails/icons/frisbeeLogo.png'
 import PcLogo from '../../../thumbnails/icons/pcLogo.jpg'
 import { type FC } from 'react'
-import { H2, TitleWrapper, AboutMeWrapper } from './styles'
+import { AboutMeTitle, TitleWrapper, AboutMeWrapper } from './styles'
 
 export const AboutMe: FC = () => {
   const about = [
@@ -17,7 +17,7 @@ export const AboutMe: FC = () => {
       <Card>
         <TitleWrapper>
           <Icon isSmall imgUrl={PcLogo} alt="PcLogo" />
-          <H2>About me</H2>
+          <AboutMeTitle>About me</AboutMeTitle>
           <Icon isSmall imgUrl={frisbeeLogo} alt="frisbeeLogo" />
         </TitleWrapper>
         {about.map((aboutme, idx) => (
