+++ src/components/general/Content/styles.tsx
